Change default cache file location to `node_modules/.cache/prettier`
