+++ src/cli/format.js
@@ -15,8 +15,8 @@ const { createIgnorer, errors } = require("./prettier-internal.js");
 const { expandPatterns, fixWindowsSlashes } = require("./expand-patterns.js");
 const getOptionsForFile = require("./options/get-options-for-file.js");
 const isTTY = require("./is-tty.js");
+const findCacheFile = require("./find-cache-file.js");
 const FormatResultsCache = require("./format-results-cache.js");
-const { createHash } = require("./utils.js");
 
 function diff(a, b) {
   return require("diff").createTwoFilesPatch("", "", a, b, "", "", {
@@ -285,38 +285,6 @@ async function formatStdin(context) {
   }
 }
 
-/**
- * Inspired by `getCacheFile` function from ESLint
- *   https://github.com/eslint/eslint/blob/c2d0a830754b6099a3325e6d3348c3ba983a677a/lib/cli-engine/cli-engine.js#L424-L485
- *
- * @param {string} cacheLocation
- * @returns {string}
- */
-async function getCacheFile(cacheLocation) {
-  const cwd = process.cwd();
-  const normalizedCacheLocation = path.normalize(cacheLocation);
-  const looksLikeADirectory = normalizedCacheLocation.slice(-1) === path.sep;
-  const resolvedCacheLocation = path.resolve(cwd, normalizedCacheLocation);
-  const getCacheFileForDirectory = () =>
-    path.join(resolvedCacheLocation, `.cache_${createHash(cwd)}`);
-  let fileStats;
-  try {
-    fileStats = await fs.lstat(resolvedCacheLocation);
-  } catch {
-    fileStats = null;
-  }
-  if (fileStats) {
-    if (fileStats.isDirectory() || looksLikeADirectory) {
-      return getCacheFileForDirectory();
-    }
-    return resolvedCacheLocation;
-  }
-  if (looksLikeADirectory) {
-    return getCacheFileForDirectory();
-  }
-  return resolvedCacheLocation;
-}
-
 async function formatFiles(context) {
   // The ignorer will be used to filter file paths after the glob is checked,
   // before any files are actually written
@@ -330,7 +298,7 @@ async function formatFiles(context) {
   }
 
   const formatResultsCache = context.argv.cache
-    ? new FormatResultsCache(await getCacheFile(context.argv.cacheLocation))
+    ? new FormatResultsCache(await findCacheFile(context.argv.cacheLocation))
     : undefined;
 
   for await (const pathOrError of expandPatterns(context)) {
